pyproject.toml
[project]
name = "nyay-sathi"
version = "2.0.0"
description = "AI-powered Indian Legal Assistant using RAG"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Abhash Chakraborty" }]
keywords = ["legal", "rag", "india", "law", "ai", "fastapi", "llm"]

dependencies = [
    # Core web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    
    # HTTP client
    "httpx>=0.28.0",
    
    # ML / Embeddings
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",
    
    # Vector search
    "faiss-cpu>=1.9.0",
    "numpy>=2.0.0",
    
    # LLM
    "groq>=0.15.0",
    
    # CLI
    "rich>=13.0.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "duckduckgo-search>=7.0.0",
]

[project.optional-dependencies]
browser = [
    # Optional: for advanced web scraping
    "playwright>=1.40.0",
]
dev = [
    "pytest>=7.0.0",
    "mypy>=1.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
nyay-cli = "cli.nyay_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "cli"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
